LayoutTests/editing/pasteboard/preserve-line-break-at-end-of-pasted-content-expected.txt

This tests for a bug where newlines would not be preserved during copy/paste. Below you should see two paragraphs containing "Hello World!" and an empty third paragraph with the caret in it.

Before copy and paste:
| "<#selection-anchor>Hello World!"
| <div>
|   id="div"
|   <#selection-focus>
|   <br>

After copy and paste:
| "Hello World!"
| <div>
|   id="div"
|   "Hello World!"
|   <br>
| <div>
|   <#selection-caret>
|   <br>
